Revert jenkins.version to and add java.version
Changed the jenkins.version to an older LTS line to get the plugin more
current.  The 1.480.3 line is no longer found in the tagged releases in
GitHub, so it'd be incredibly difficult to diagnose any core issues
found with that version.  Also, I kept the java version at 6 to match
the requirement at 1.609.3.