[JENKINS-35283]: Update to parent POM
* Update parent POM

Updated to the parent pom and the Jenkins version to 1.625.3.  There were some changes in the findbugs-maven-plugin settings, but running tests with these changes didn't produce any failures.

* Fix findbugs errors

* Revert jenkins.version to and add java.version

Changed the jenkins.version to an older LTS line to get the plugin more current.  The 1.480.3 line is no longer found in the tagged releases in GitHub, so it'd be incredibly difficult to diagnose any core issues found with that version.  Also, I kept the java version at 6 to match the requirement at 1.609.3.

* Security fix for INFRA-588

Change `http` to `https`

* Fixes to various ants.